massage Travis configuration some more

+++ deploy-examples.sh
@@ -14,9 +14,12 @@ sysconfcpus/bin/sysconfcpus -n $CPUNUM elm-make Example3b.elm --output out/Examp
 sysconfcpus/bin/sysconfcpus -n $CPUNUM elm-make Example4.elm --output out/Example4.html --yes
 cd out
 
-git init
-git config user.name "Travis CI"
-git config user.email "jvoigtlaender@users.noreply.github.com"
-git add .
-git commit -m "Travis deploy examples to gh-pages"
-git push --force --quiet "https://$GH_TOKEN@github.com/$TRAVIS_REPO_SLUG" master:gh-pages >/dev/null 2>&1
+if [[ "$TRAVIS_PULL_REQUEST" == "false" && "$TRAVIS_BRANCH" == "master" ]];
+then
+  git init --quiet;
+  git config user.name "Travis CI";
+  git config user.email "jvoigtlaender@users.noreply.github.com";
+  git add .;
+  git commit -m "Travis deploy examples to gh-pages";
+  git push --force --quiet "https://$GH_TOKEN@github.com/$TRAVIS_REPO_SLUG" master:gh-pages >/dev/null 2>&1;
+fi
